Move KonashiListener to listeners package
