+++ Konashi/konashi-v2-android-sdk/src/main/java/com/uxxu/konashi/lib/KonashiApiInterface.java
@@ -3,6 +3,8 @@ package com.uxxu.konashi.lib;
 import android.app.Activity;
 import android.content.Context;
 
+import com.uxxu.konashi.lib.listeners.KonashiListener;
+
 /**
  * konashi APIのインタフェース
  *
