+++ Konashi/konashi-v2-android-sdk/src/main/java/com/uxxu/konashi/lib/KonashiManager.java
@@ -2,7 +2,8 @@ package com.uxxu.konashi.lib;
 
 import android.content.Context;
 
-import java.util.ArrayList;
+import com.uxxu.konashi.lib.listeners.KonashiListener;
+
 import java.util.List;
 
 /**
