+++ Konashi/konashi-v2-android-sdk/src/main/java/com/uxxu/konashi/lib/listeners/KonashiListener.java
@@ -1,4 +1,4 @@
-package com.uxxu.konashi.lib;
+package com.uxxu.konashi.lib.listeners;
 
 import com.uxxu.konashi.lib.listeners.KonashiAnalogListener;
 import com.uxxu.konashi.lib.listeners.KonashiConnectionListener;
