+++ apps/app/src/main/java/com/e10dokup/konashi_sdk_test/MainActivity.java
@@ -11,7 +11,7 @@ import android.widget.Toast;
 
 import com.uxxu.konashi.lib.Konashi;
 import com.uxxu.konashi.lib.KonashiErrorReason;
-import com.uxxu.konashi.lib.KonashiListener;
+import com.uxxu.konashi.lib.listeners.KonashiListener;
 import com.uxxu.konashi.lib.ui.KonashiActivity;
